% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netShiny.R
\name{netShiny}
\alias{netShiny}
\title{This function launches the netShiny Shiny app.}
\usage{
netShiny(Net.obj = NULL, mapping = NULL, resamples = NULL)
}
\arguments{
\item{Net.obj}{A list containing correlation matrices and/or dataframes.}

\item{mapping}{A dataframe containing the mapping of the nodes in the networks. The names must match the column names in the Net.obj argument.}

\item{resamples}{A list containing a list of resampling data of the Net.obj argument.}
}
\value{
A Shiny app.
}
\description{
This function launches the netShiny Shiny app.
}
\examples{
netShiny()
}
\references{
Behrouzi, P., and Wit, E. C. (2017c). netgwas: An R Package for Network-Based Genome-Wide Association Studies. arXiv preprint, arXiv:1710.01236.
}
\seealso{
\code{\link[netgwas]{netphenogeno}}, \code{\link[netgwas]{selectnet}}
}
\author{
Rocherno de Jongh and Pariya Behrouzi \cr
Maintainer: Rocherno de Jongh \email{rocherno.dejongh@hotmail.com}
}
